Minor tweak to P_RandomHitscanAngle()

+++ src/m_random.c
@@ -52,7 +52,7 @@ unsigned int    bigseed;
 //
 int P_RandomHitscanAngle(fixed_t spread)
 {
-    return (int)(((spread < 0 ? FixedToAngle(-spread) : FixedToAngle(spread)) * M_SubRandom()) / 255);
+    return (int)(((int64_t)(spread < 0 ? FixedToAngle(-spread) : FixedToAngle(spread)) * M_SubRandom()) / 255);
 }
 
 //
